Hoist engineered_features to BotorchSurrogate; share get_feature_indices (#5)
Per reviewer discussion, avoid duplicating `get_feature_indices`. It needs
`inputs`, `categorical_encodings` and `engineered_features`; the first two
are already on `BotorchSurrogate`, so `engineered_features` is hoisted to
join them:

- `engineered_features` field + `validate_aggregations` move from
  `TrainableSurrogate` to the `BotorchSurrogate` data model; the duplicate
  field declaration on `LinearDeterministicSurrogate` is removed.
- functional `BotorchSurrogate.__init__` sets `engineered_features`;
  redundant assignments drop from `TrainableBotorchSurrogate`,
  `LinearDeterministicSurrogate` and `PairwiseGPSurrogate`.
- `get_feature_indices` moves to functional `BotorchSurrogate`; the copies
  on `TrainableBotorchSurrogate` and the `PairwiseTrainableSurrogate` mixin
  are removed. `SingleTaskGP` and `PairwiseGP` now inherit one shared copy.

This supersedes the mixin-local copy from commit bb4d1fda. The
`CategoricalDeterministicSurrogate` spec gains the inherited
`engineered_features` field.
